feat!(storage): use Django's storage backend for object storage
For backward compatibility and per object file handling before calling
the storage backend the `FileField` is extended to dynamically
select the storage backend based on the `encryption_status` of the
`File` instance.

Overriding `get_object_parameters` could be done in place in the
`DynamicFileField`. Loading a class is however more modular and
allows for environment based configuration.