chore: add some comments so we don't forget factoring out storage code
We decided to (soon) factor out the file storage code into it's own
proposed module / package, so it can be reused outside of Alexandria.

To ensure we don't forget it next time we work on this code, we're adding
a few TODO lines in context where it may make sense.