fix(upload): use content-type and filename to infer mime types
This fixes an issue we saw when uploading docx documents generated by
document-merge-service. We were previously relying solely on
content-based mime type detection (using python-magic), which caused an
incorrect detection for such documents as ZIP files instead of docx.
This change makes use of the type information we have available:

1. Content-type header
2. If it's application/octet-stream, guess mime type from file extension
3. If no guess is available, content-based guessing based (as before)