feat(document)!: create document and file in one request!
BREAKING CHANGE:
The document post endpoint now requires the file data to be provided
as well.
The reason for this change is allowing the frontend to create documents
and files in one request, preventing documents with no associated files.
Which fixes the problem if the file got rejected for any reason, the
application would create an empty document.